+++ integration-tests/environments/react-native/harness/xcode-cli.js
@@ -53,6 +53,28 @@ function list(type, searchterm, asJson = true) {
 }
 
 /**
+ * Get My Mac Device ID
+ * @param
+ */
+function getMyMacDeviceId() {
+  const p = xcrun("xctrace", "list", "devices");
+
+  // The machine running this command will always be displayed as the second item
+  const outputLines = p.stdout.split("\n");
+  const myMac = outputLines[1];
+  const idMatcher = /\(([\S]+)\)/;
+  const searchResult = myMac.match(idMatcher);
+
+  if (searchResult) {
+    const myMacDeviceId = searchResult[1];
+    console.log(myMacDeviceId);
+    return myMacDeviceId;
+  }
+
+  return null;
+}
+
+/**
  * Create a new device
  * @param {string} name The name of the new device
  * @param {string} deviceTypeId The type of device to use (run `xcrun simctl list devicetypes` to get these)
@@ -130,6 +152,7 @@ function bootstatus(device) {
 }
 
 module.exports = {
+  getMyMacDeviceId,
   xcrun,
   simctl: {
     list,